Refactored source processing so that we can use UMD builds
This is pretty much the biggest hack we do. Loader.js as used by Ember doesn't declare itself as a true AMD loader (don't get me started). So we trick it by defining AMD for the UMD modules as declared by our d3 packages. This is implemented as a brocolli filter.